Uses okio/moshi natively in Elasticsearch everywhere
Before, we used a combination of our buffer and Okio's Buffer, and a
combination of our json encoder and Moshi's. This changes all the code
in Elasticsearch to use Okio buffers which reduces the amount of array
allocations and allows us to do cheap content hashes in the future.

The content hashes will be used to obviate duplicate document indexing,
for example, when reporters backoff and retry with the same span. This
will help simplify analysis.